web/requirements.txt
django-filter
djangorestframework
asgiref==3.2.3
Django==3.0.3
numpy==1.18.1
pandas==1.0.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
